Update dynamic headline component
